requirements.txt
Flask==2.2.3
Flask_SocketIO==5.3.3
python-dotenv==1.0.0
requests==2.28.1
gevent==21.12.0
gevent-websocket==0.10.1

